# public/openapi.yaml
openapi: 3.1.0
info:
  title: GroLabs Public API
  version: 1.0.0
  summary: Search, analytics events, and storefront diagnostics for any platform.
  description: |
    The GroLabs public API. The mental model is **Meilisearch (search) + PostHog
    (analytics)**, pointed at GroLabs so every storefront interaction is captured.

    There are three trust models on this surface:

    - **Origin-validated (no key):** `/search`, `/search/token`, `/events`,
      `/events/token`. The browser sends no secret. The request's `Origin` header
      is matched server-side against the instance's registered
      `storefront_domains`. `instance_id` is public — treat it like a Stripe
      *publishable* key. Only registered domains can call these.
    - **Write-key (Bearer):** `/catalog/*`. Privileged, server-side only. The
      per-instance write key is secret — never ship it to a browser. Treat it
      like a Stripe *secret* key.
    - **Anonymous public:** `/diagnostic/*`. No instance, no key; IP rate-limited.
      The `runId` UUID doubles as an unguessable share token.

    This document is the single source of truth for the API contract. The
    hand-written JavaScript SDK (`@grolabs/web-sdk`) and every generated
    per-language SDK derive from it.

    > **Spec status — v1.0.0.** Covers the full external-platform surface:
    > search, events, tenant tokens, catalog **document ingestion**
    > (`/catalog/documents`, `/catalog/documents/delete`, `/catalog/tasks`,
...

  contact:
...
servers:
...
security: []   # Most endpoints are origin-validated, not credential-authenticated.
paths:
...
  /catalog/documents:
    post:
      tags: [Catalog]
      summary: Upsert catalog documents
      operationId: upsertDocuments
      security:
        - writeKey: []
      description: |
        Add-or-update a batch of catalog documents (max 1000 per request). The
        first full load is just a large upsert. Documents are mapped into the
        GroLabs canonical structure for ranking/filtering and the original
        payload is retained verbatim for display. Accept-fast — poll the returned
        `task_id` for terminal status.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentsUpsertRequest' }
      responses:
        '202':
          description: Batch accepted (indexing is async).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestAck' }
        '400':
          description: Invalid instance_id, invalid JSON, empty/oversized batch, or no documents.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Missing or invalid write key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Ingestion failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Catalog]
      summary: Delete the entire catalog
      operationId: deleteAllDocuments
      security:
        - writeKey: []
      description: Remove every document for the instance (a re-seed). Accept-fast.
      parameters:
        - name: instance_id
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '202':
          description: Deletion accepted (async).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestAck' }
        '400':
          description: Invalid instance_id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Missing or invalid write key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Deletion failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /catalog/tasks/{taskId}:
    get:
      tags: [Catalog]
      summary: Poll an ingestion task
      operationId: getCatalogTask
      security:
        - writeKey: []
      description: |
        Read the status of an ingestion task. A still-processing task is refreshed
        against the search engine and converges to `succeeded`/`failed` on read.
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: instance_id
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: The task status.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskStatus' }
        '401':
          description: Missing or invalid write key.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Task not found for this instance.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: Task read failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    writeKey:
      type: http
      scheme: bearer
      description: |
        Per-instance write key, sent as `Authorization: Bearer <write_key>`.
        Secret — server-side only. Never expose in a browser.
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Stable machine-readable error code.
          examples: [instance_not_found_or_origin_not_authorized, rate_limited, url_required]
        message:
          type: string
          description: Optional human-readable detail (not always present).
...
    DocumentsUpsertRequest:
      type: object
      required: [instance_id, documents]
      properties:
        instance_id:
          oneOf:
            - { type: integer, minimum: 0 }
            - { type: string }
        documents:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            type: object
            description: |
              An arbitrary merchant document. Recognized canonical fields:
              `id` (required — your stable primary key), `title`/`name`,
              `description`, `brand`, `categories`, `category_ids`, `price`,
              `in_stock`, `image`/`image_url`, `url`, `variants`. Any other field
              is stored and returned verbatim on search.
            properties:
              id:
                oneOf: [{ type: string }, { type: integer }]
            additionalProperties: true
        primary_key:
          type: string
          default: id
          description: The field to use as the document id. Defaults to `id`.
...
    IngestAck:
      type: object
      required: [task_id, meilisearch_task_uid]
      properties:
        task_id:
          oneOf: [{ type: string, format: uuid }, { type: 'null' }]
          description: GroLabs ingestion task id — poll `/catalog/tasks/{task_id}`.
        accepted:
          type: integer
          description: Documents accepted for indexing.
        rejected:
          type: integer
          description: Documents rejected (e.g. missing primary key).
        meilisearch_task_uid:
          oneOf: [{ type: integer }, { type: 'null' }]
    TaskStatus:
      type: object
      required: [task_id, op, status, document_count, failed_count, created_at]
      properties:
        task_id: { type: string, format: uuid }
        op:
          type: string
          enum: [upsert, delete, delete_all, settings]
        status:
          type: string
          enum: [queued, processing, succeeded, failed]
        document_count: { type: integer }
        failed_count: { type: integer }
        meilisearch_task_uid:
          oneOf: [{ type: integer }, { type: 'null' }]
        error:
          description: Error detail when the task failed; null otherwise.
          oneOf: [{ type: object, additionalProperties: true }, { type: 'null' }]
        created_at: { type: string, format: date-time }
        completed_at:
          oneOf: [{ type: string, format: date-time }, { type: 'null' }]
...